Move archive related code back to util.go
